Deduplicate `spksndtype` and `SPKRtable` structs
