+++ src/spksndtype.rs
@@ -1,0 +1,8 @@
+#[derive(Copy, Clone)]
+#[repr(C, packed)]
+pub struct spksndtype {
+    pub start: u16,
+    pub priority: u8,
+    pub samplerate: u8,
+    pub name: [i8; 12],
+}
